main -> feed로 변경, viewmodel 수정
